requirements.txt
Django==2.2.6
djangorestframework==3.10.3
psycopg2==2.8.4
django-mptt==0.10.0
pika==1.1.0
requests==2.22.0
PyYAML==5.1.2
pandas==0.25.2